renamed fit submodule to optimize
